run/requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google.cloud.storage==1.44.0
qrcode[pil]==7.3.1
matplotlib
google.cloud.bigquery